requirements.txt
archspec @ file:///croot/archspec_1697725767277/work
biopython==1.84
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///work/perseverance-python-buildout/croot/brotli-split_1698805593785/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
conda @ file:///croot/conda_1708369113911/work
conda-content-trust @ file:///work/perseverance-python-buildout/croot/conda-content-trust_1698882886606/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1702997573971/work/src
conda-package-handling @ file:///work/perseverance-python-buildout/croot/conda-package-handling_1698851267218/work
conda_package_streaming @ file:///work/perseverance-python-buildout/croot/conda-package-streaming_1698847176583/work
contourpy==1.3.0
cryptography @ file:///croot/cryptography_1702070282333/work
cycler==0.12.1
distro @ file:///work/perseverance-python-buildout/croot/distro_1701732366176/work
fonttools==4.54.1
idna @ file:///work/perseverance-python-buildout/croot/idna_1698845632828/work
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
kiwisolver==1.4.7
libmambapy @ file:///work/perseverance-python-buildout/croot/mamba-split_1701744133524/work/libmambapy
logomaker==0.8
matplotlib==3.9.2
menuinst @ file:///croot/menuinst_1706732933928/work
numpy==2.1.2
packaging @ file:///work/perseverance-python-buildout/croot/packaging_1698805390494/work
pandas==2.2.3
pillow==11.0.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
pycosat @ file:///work/perseverance-python-buildout/croot/pycosat_1698863456259/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyparsing==3.2.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
pytz==2024.2
requests @ file:///croot/requests_1707355572290/work
ruamel.yaml @ file:///work/perseverance-python-buildout/croot/ruamel.yaml_1698863605521/work
scipy==1.14.1
seaborn==0.13.2
setuptools==68.2.2
six==1.16.0
tqdm @ file:///work/perseverance-python-buildout/croot/tqdm_1701735729845/work
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
tzdata==2024.2
urllib3 @ file:///croot/urllib3_1707770551213/work
wheel==0.41.2
zstandard @ file:///work/perseverance-python-buildout/croot/zstandard_1698847073368/work
